+++ src/Homepage/HomePage.jsx
@@ -1,0 +1,15 @@
+import HomePageHero from './HomePageHero';
+import HomePageAbout from './HomePageAbout';
+import HomePageServices from './HomePageServices';
+
+const HomePage = () => {
+  return (
+    <>
+      <HomePageHero />
+      <HomePageAbout />
+      <HomePageServices />
+    </>
+  );
+};
+
+export default HomePage;
